refactor: 임시 저장 관련 DTO 묶어서 관리 - #63
