+++ src/main/java/kr/hanjari/backend/web/dto/club/response/draft/ClubDetailDraftResponseDTO.java
@@ -1,4 +1,4 @@
-package kr.hanjari.backend.web.dto.club.response;
+package kr.hanjari.backend.web.dto.club.response.draft;
 
 import kr.hanjari.backend.domain.draft.ClubDetailDraft;
 import kr.hanjari.backend.domain.enums.RecruitmentStatus;
